requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.12.14
charset-normalizer==3.4.0
dateparser==1.2.0
frozenlist==1.5.0
idna==3.10
joblib==1.4.2
multidict==6.1.0
numpy==1.21.2
pandas==2.0.3
propcache==0.2.1
pycryptodome==3.21.0
python-binance==1.0.25
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.0.2
scipy==1.10.1
six==1.17.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
websockets==14.1
xgboost==1.7.6
yarl==1.18.3
